Improved apps installation

+++ app/src/main/java/ru/playsoftware/j2meloader/util/FileUtils.java
@@ -21,32 +21,29 @@ import android.content.Context;
 import android.net.Uri;
 import android.util.Log;
 
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.nio.channels.FileChannel;
 import java.util.Arrays;
-import java.util.Iterator;
-import java.util.LinkedHashMap;
-import java.util.Map;
 
 public class FileUtils {
 
 	private static String TAG = FileUtils.class.getName();
+	private static final String TEMP_JAR_NAME = "tmp.jar";
+	private static final String TEMP_JAD_NAME = "tmp.jad";
 	private static final int BUFFER_SIZE = 1024;
 
 	public static void copyFiles(String src, String dest, FilenameFilter filter) {
-		File fsrc = new File(src);
-		File fdest = new File(dest);
-		fdest.mkdirs();
+		File srcFile = new File(src);
+		File dstFile = new File(dest);
+		dstFile.mkdirs();
 		String to;
-		File[] list = fsrc.listFiles(filter);
+		File[] list = srcFile.listFiles(filter);
 		for (File entry : list) {
 			to = entry.getPath().replace(src, dest);
 			if (entry.isDirectory()) {
@@ -70,55 +67,36 @@ public class FileUtils {
 	public static void deleteDirectory(File dir) {
 		if (dir.isDirectory()) {
 			File[] listFiles = dir.listFiles();
-			if (listFiles.length != 0) {
+			if (listFiles != null && listFiles.length != 0) {
 				for (File file : listFiles) {
 					deleteDirectory(file);
 				}
 			}
 		}
-		dir.delete();
-	}
-
-	public static LinkedHashMap<String, String> loadManifest(File mf) {
-		LinkedHashMap<String, String> params = new LinkedHashMap<>();
-		try {
-			BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream(mf)));
-			String line;
-			int index;
-			while ((line = br.readLine()) != null) {
-				index = line.indexOf(':');
-				if (index > 0) {
-					params.put(line.substring(0, index).trim(), line.substring(index + 1).trim());
-				}
-				if (line.length() > 0 && Character.isWhitespace(line.charAt(0))) {
-					Iterator<Map.Entry<String, String>> iter = params.entrySet().iterator();
-					Map.Entry<String, String> entry = null;
-					while (iter.hasNext()) {
-						entry = iter.next();
-					}
-					params.put(entry.getKey(), entry.getValue() + line.substring(1));
-				}
-			}
-			br.close();
-		} catch (Throwable t) {
-			Log.e(TAG, "getAppProperty() will not be available due to " + t.toString());
+		if (!dir.delete()) {
+			Log.w(TAG, "Can't delete file: " + dir);
 		}
-		return params;
 	}
 
 	public static String getAppPath(Context context, Uri uri) throws IOException {
+		if ("file".equals(uri.getScheme())) {
+			String path = uri.getPath();
+			if (path != null && new File(path).exists()) {
+				return path;
+			}
+		}
 		InputStream in = context.getContentResolver().openInputStream(uri);
 		OutputStream out = null;
-		File folder = new File(context.getApplicationInfo().dataDir, JarConverter.TEMP_URI_FOLDER_NAME);
-		folder.mkdir();
+		File folder = context.getCacheDir();
 		byte[] signature = new byte[2];
 		byte[] jarSignature = new byte[]{0x50, 0x4B};
-		in.read(signature);
+		if (in == null || in.read(signature) == -1)
+			throw new IOException("Can't read data from uri: " + uri);
 		File file;
 		if (Arrays.equals(signature, jarSignature)) {
-			file = new File(folder, JarConverter.TEMP_JAR_NAME);
+			file = new File(folder, TEMP_JAR_NAME);
 		} else {
-			file = new File(folder, JarConverter.TEMP_JAD_NAME);
+			file = new File(folder, TEMP_JAD_NAME);
 		}
 		try {
 			out = new FileOutputStream(file);
